tdl-api/requirements.txt
fastapi==0.67.0
uvicorn
tortoise-orm
tortoise-orm[asyncpg]